remove pipeline directory to types + utils
