+++ frontend/interactEM/src/utils/pipeline.ts
@@ -1,63 +1,16 @@
 import { type Edge, type Node, Position, type XYPosition } from "@xyflow/react"
-
-import type { OperatorParameter, OperatorTag } from "../client"
-
 import { edgeOptions } from "../components/composerpipelineflow"
-import type { NodeType } from "../types/nodes"
+import type {
+  EdgeJSON,
+  OperatorJSON,
+  OperatorNodeData,
+  PipelineJSON,
+  PortJSON,
+} from "../types/pipeline"
 
 const position: XYPosition = {
   x: 0,
   y: 0,
-}
-
-export type OperatorNodeData = {
-  label: string
-  image: string
-  inputs?: string[]
-  outputs?: string[]
-  parameters?: OperatorParameter[]
-  tags?: OperatorTag[]
-  type: NodeType
-}
-
-export interface OperatorJSON extends OperatorNodeData {
-  id: string
-
-  [x: string | number | symbol]: unknown
-}
-
-export interface EdgeJSON {
-  input_id: string
-  output_id: string
-
-  [x: string | number | symbol]: unknown
-}
-
-type PortNodeType = "port"
-
-export interface PortJSON {
-  node_type: PortNodeType
-  id: string
-  operator_id: string
-  portkey: string
-  port_type: string
-
-  [x: string | number | symbol]: unknown
-}
-
-interface PipelineData {
-  operators: OperatorJSON[]
-  ports: PortJSON[]
-  edges: EdgeJSON[]
-}
-
-export interface PipelineJSON {
-  data: PipelineData
-}
-
-export interface PipelineWithID {
-  id: string
-  data: PipelineJSON["data"]
 }
 
 export const fromPipelineJSON = (pipelineJSON: PipelineJSON) => {
